# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "JWE JWS Encryption Rest Service",
    "version": "Version 1.0.0",
    "title": "JWE JWS Encryption Rest Service",
...
  },
  "basePath": "/rest/api/",
...
  "paths": {
    "/createJweUsingRsaPki": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Create JWE using RSA PKI",
        "operationId": "createJweUsingRsaPkiUsingPOST",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "payload",
            "description": "Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "vEncryptionKid",
            "description": "Visa Encryption Kid (Cert ID)",
            "in": "formData",
            "required": true,
            "type": "string",
            "default": "83F4EEB2"
          },
          {
            "name": "vEncryptionCertificateFile",
            "description": "Visa Encryption Certificate Public Key (PEM Format) File",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "vEncryptionCertificatePem",
            "description": "Visa Encryption Certificate Public Key (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateJweResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/decryptJweUsingRsaPki": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Decrypt JWE Using RSA PKI",
        "operationId": "decryptJweUsingRsaPkiUsingPOST",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jwe",
            "description": "JWE Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "privateKeyFile",
            "description": "Private Key File (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "privateKeyPem",
            "description": "Private Key (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateJweResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/createJwsUsingRsaPki": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Sign the JWE Using RSA PKI",
        "operationId": "createJwsUsingRsaPkiUsingPost",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jwe",
            "description": "JWE Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "signingKid",
            "description": "Signing Kid",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "privateKeyFile",
            "description": "Private Key File (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "privateKeyPem",
            "description": "Private Key (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateJwsResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/verifyJwsUsingRsaPki": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Verify the JWS Using RSA PKI",
        "operationId": "verifyJwsUsingRsaPkiUsingPost",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jws",
            "description": "JWS Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "signingCertificateFile",
            "description": "Signing Certificate Public Key (PEM Format) File",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "signingCertificatePem",
            "description": "Signing Certificate Public Key (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/VerifyJwsResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/createJweJwsUsingRsaPki": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Create JWE & JWS using RSA PKI",
        "operationId": "createJweJwsUsingRsaPkiUsingPOST",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "payload",
            "description": "Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "vEncryptionKid",
            "description": "Visa Encryption Kid (Cert ID)",
            "in": "formData",
            "required": true,
            "type": "string",
            "default": "83F4EEB2"
          },
          {
            "name": "vEncryptionCertificateFile",
            "description": "Visa Encryption Certificate Public Key (PEM Format) File",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "vEncryptionCertificatePem",
            "description": "Visa Encryption Certificate Public Key (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "clientSigningKid",
            "description": "Client Signing Kid",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "clientSigningPrivateKeyFile",
            "description": "Client Signing Private Key File (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "clientSigningPrivateKeyPem",
            "description": "Client Signing Private Key (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateJweResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/verifyAndDecryptUsingRsaPki": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Verify JWS & Decrypt Extracted JWE using RSA PKI",
        "operationId": "verifyAndDecryptUsingRsaPkiUsingPOST",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jws",
            "description": "JWS Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "vSigningCertificatePem",
            "description": "Visa Signing Certificate Public Key (PEM Format)",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "clientPrivateKeyPem",
            "description": "Client Private Key (PEM Format)",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DecryptJweResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/createJwe": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Create JWE Using API Key & Shared Secret",
        "operationId": "createJweUsingPOST",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "payload",
            "description": "Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "apiKey",
            "description": "Your Encryption API Key",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "sharedSecret",
            "description": "Your Encryption Shared Secret",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateJweResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/decryptJwe": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Decrypt JWE Using Shared Secret",
        "operationId": "decryptJweUsingPOST",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jwe",
            "description": "JWE Payload",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "sharedSecret",
            "description": "Your Encryption Shared Secret",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateJweResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/signBindingRequest": {
      "post": {
        "tags": [
          "jwe-jws-encryption-rest-service"
        ],
        "summary": "Sign Device Binding Request",
        "operationId": "signBindingRequestUsingPOST",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "clientDeviceID",
            "description": "Client Device ID",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "clientReferenceID",
            "description": "Client Reference ID",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "vProvisionedTokenID",
            "description": "Provisioned Token ID",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "nonce",
            "description": "Nonce",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "signingKid",
            "description": "Signing Kid",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "privateKeyFile",
            "description": "Private Key File (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "privateKeyPem",
            "description": "Private Key (PEM Format)",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateJweResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "CreateJweResponse": {
      "type": "object",
      "properties": {
        "encData": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CreateJweResponse"
    },
    "CreateJwsResponse": {
      "type": "object",
      "properties": {
        "encData": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CreateJwsResponse"
    },
    "VerifyJwsResponse": {
      "type": "object",
      "properties": {
        "verified": {
          "type": "boolean"
        },
        "payload": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "VerifyJwsResponse"
    },
    "DecryptJweResponse": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "DecryptJwsResponse"
    }
  }
}